refactor(logic): split order_engine validation phases into part file (#3290)
Extract per-category validation helpers, run state, and phase runners
from the 732-line order_engine.dart god-file into order_engine_validation.dart
as a `part of` fragment. The library file keeps slot wiring, generated mixin
integration, and the public OrderEngine surface; behaviour is unchanged.

Refs #3290
